src/data-repository/models/ticket.ts
export interface Ticket {
  _id: string;
  url: string;
  external_id: string;
  created_at: string;
  type?: string;
  subject: string;
  description?: string;
  priority: string;
  status: string;
  submitter_id: number;
  assignee_id?: number;
  organization_id?: number;
  tags: string[];
  has_incidents: boolean;
  due_at?: string;
  via: string;
}

export const getTicket = (props: Ticket) => ({
  ...props,
  getSubmitter: () => {},
  getAssignee: () => {},
  getOrganization: () => {},
  getSearchableFields: () => Object.keys(props),
});
